Update Exercise3.java
Make Hashcode method to return id only so that if you add p1 for first time, the content will be placed in bucket with hascode 1. When you add it to second time since hashcode is same, it goes for same bucket and check for equals and equals will return true so it will override p1.