Update Exercise3.java
Hashcode method should return hash code value of Person id, so that if you add Person p1 for first time, the object p1 will be stored in bucket array with hashcode value. 
When we add it for the second time since hashcode value is same, it goes for same bucket location and check for equals. 
If equals return true, it will override the old p1 object.