Interview Test
I wrote in what I would fix for the sample application, and wrote a unit test for the unit test app. I saw that we could refactor the application, but I did not actually refactor instead I wrote the unit test based on the following assumptions: 1. Each repository has an interface it implements so it can be easily mocked 2. The function returns a lesson when done so you can assert from that.